backend 40%

+++ app/Http/Controllers/API/AuthenticationController.php
@@ -108,6 +108,14 @@ class AuthenticationController extends Controller
             }
 
             $user = User::where('email', $credentials['email'])->first();
+            
+            if (!$user->email_verified_at) {
+                return response()->json([
+                    'response_code' => 401,
+                    'status'        => 'error',
+                    'message'       => 'Email not verified',
+                ], 401);
+            }
             $user->tokens()->delete();
             $token = $user->createToken('admin-auth-token')->plainTextToken;
